+++ arkenv/src/main/kotlin/com/apurebase/arkenv/ArkenvExceptions.kt
@@ -25,3 +25,17 @@ internal class UnsupportedMappingException(key: String, clazz: KClass<*>) : Arke
 internal class MappingException(key: String, value: String, clazz: KClass<*>, cause: Exception) : ArkenvException(
     "Could not parse property '$key' with value '$value' as class '$clazz'", cause
 )
+
+/**
+ * Unchecked exception thrown when no value can be found for the given name.
+ */
+internal class MissingArgumentException(name: String, info: String): ArkenvException(
+    "No value passed for property $name ($info)"
+)
+
+/**
+ * Unchecked exception thrown when the requested feature could not be found.
+ */
+internal class FeatureNotFoundException(featureName: String?) : ArkenvException(
+    "Feature $featureName could not be found. Make sure it was installed correctly."
+)