+++ arkenv/src/main/kotlin/com/apurebase/arkenv/feature/PlaceholderParser.kt
@@ -2,7 +2,7 @@ package com.apurebase.arkenv.feature
 
 import com.apurebase.arkenv.ArgumentDelegate
 import com.apurebase.arkenv.Arkenv
-import com.apurebase.arkenv.feature.EnvironmentVariableFeature.Companion.getEnv
+import com.apurebase.arkenv.MissingArgumentException
 import com.apurebase.arkenv.toSnakeCase
 
 /**
@@ -44,9 +44,8 @@ internal class PlaceholderParser : ProcessorFeature {
     private fun findPlaceholderReplacement(placeholder: String): String =
         findReplacementInDelegates(arkenv.delegates, placeholder)
                 ?: arkenv.getOrNull(placeholder)
-                ?: getEnv(placeholder, false)
                 ?: findReplacementInArgs(arkenv.argList, placeholder)
-                ?: throw IllegalArgumentException("Cannot find value for placeholder $placeholder")
+                ?: throw MissingArgumentException(placeholder, "Cannot find value for placeholder")
 
     private fun findReplacementInDelegates(delegates: Collection<ArgumentDelegate<*>>, placeholder: String): String? =
         delegates