+++ arkenv/src/main/kotlin/com/apurebase/arkenv/feature/ProfileFeature.kt
@@ -1,33 +1,22 @@
 package com.apurebase.arkenv.feature
 
 import com.apurebase.arkenv.Arkenv
-import com.apurebase.arkenv.DEPRECATED_GENERAL
-import com.apurebase.arkenv.argument
-import com.apurebase.arkenv.parse
+import com.apurebase.arkenv.getFeature
+import com.apurebase.arkenv.split
 
 /**
  * Feature for loading profile-based configuration.
  * A list of active profiles can be configured via the *ARKENV_PROFILE* argument.
- * @param name overrides the default name of the profile argument, can be set via *ARKENV_PROFILE*
  * @param prefix the default prefix for any profile configuration files, can be set via *ARKENV_PROFILE_PREFIX*
  * @param locations defines the default list of locations in which to look for profile configuration files,
  * can be set via *ARKENV_PROFILE_LOCATION*
  * @param parsers additional providers for profile file parsing. By default supports the property format.
  */
-class ProfileFeature
-@Deprecated(DEPRECATED_GENERAL)
-constructor(
-    name: String = "--arkenv-profile",
-    prefix: String = "application",
-    locations: Collection<String> = listOf(),
+class ProfileFeature(
+    private var prefix: String = "application",
+    private var locations: Collection<String> = listOf(),
     parsers: Collection<PropertyParser> = listOf()
-) : ArkenvFeature, Arkenv("ProfileFeature") {
-
-    constructor(
-        prefix: String = "application",
-        locations: Collection<String> = listOf(),
-        parsers: Collection<PropertyParser> = listOf()
-    ) : this("--arkenv-profile", prefix, locations, parsers)
+) : ArkenvFeature {
 
     private val parsers: MutableList<PropertyParser> = mutableListOf(::PropertyFeature)
 
@@ -35,26 +24,19 @@ constructor(
         this.parsers.addAll(parsers)
     }
 
-    internal val profiles: List<String> by argument(name) {
-        defaultValue = ::emptyList
-    }
-
-    private val prefix: String by argument("--arkenv-profile-prefix") {
-        defaultValue = { prefix }
-    }
-
-    private val location: Collection<String> by argument("--arkenv-profile-location") {
-        defaultValue = { locations }
-    }
+    var active: List<String> = listOf()
+        private set
 
     override fun onLoad(arkenv: Arkenv) {
-        parse(arkenv.argList.toTypedArray())
+        active = arkenv.getOrNull("--arkenv-profile")?.split() ?: emptyList()
+        prefix = arkenv.getOrNull("--arkenv-profile-prefix") ?: prefix
+        locations = arkenv.getOrNull("--arkenv-profile-location")?.split() ?: locations
         load(arkenv, null)
-        profiles.forEach { load(arkenv, it) }
+        active.forEach { load(arkenv, it) }
     }
 
     private fun load(arkenv: Arkenv, file: String?) = parsers
-        .map { it(makeFileName(file), location) }
+        .map { it(makeFileName(file), locations) }
         .forEach { it.onLoad(arkenv) }
 
     private fun makeFileName(profile: String?) =
@@ -63,3 +45,5 @@ constructor(
 }
 
 typealias PropertyParser = (String, Collection<String>) -> PropertyFeature
+
+val Arkenv.profiles get() = getFeature<ProfileFeature>()